Implemented controller concern for custom devise params
