+++ app/controllers/application_controller.rb
@@ -1,12 +1,3 @@
 class ApplicationController < ActionController::Base
-
-  before_action :configure_permitted_parameters,
-                if: :devise_controller?
-
-
-  def configure_permitted_parameters
-    # devise knows sign_up and account_update keywords on usage.
-    devise_parameter_sanitizer.permit(:sign_up, keys: [:nickname])
-    devise_parameter_sanitizer.permit(:account_update, keys: [:nickname])
-  end
+  include DeviseWhitelist
 end
